+++ lib/cache_network_media.dart
@@ -3,3 +3,5 @@ library;
 
 // Export main widget (public API)
 export 'src/widgets/cache_network_media_widget.dart';
+export 'src/providers/cache_network_media_image_provider.dart';
+export 'src/core/cache_network_media.dart';